requirements.txt
asgiref==3.8.1
chardet==5.2.0
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
pillow==11.0.0
playsound==1.2.2
python-decouple==3.8
reportlab==4.2.5
sqlparse==0.5.3
tzdata==2024.2
